drm/i915: Fix up vrr timings
The hardware/spec seems to riddled with off-by-one errors. Try to
deal with them all to get the timings to make sense.

With these adjustments we now get the same min vtotal as w/o vrr,
and we are able to accurately predict on which scanline the
start of vblank interrupt will happen.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>