drm/i915: Skip DPLL forcing if the pipe is already on
DSI has its own DPLL, and so let's assume we don't beed to force
the DPLL when we borrow the PPS while the pipe is in use by DSI.
For simplicity let's just assume there's a clock if the pipe is
already on.

This avoids the "pipe alreayd on" WARNS from the DPLL code.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>