drm/i915: Fix cursor coordinates on bigjoiner slave
Adjust the cursor dst coordinates appripriately when it's on
the bigjoiner slave pipe. intel_atomic_plane_check_clipping()
already did this but with the cursor we discard those results
(apart from uapi.visible and error checks) since the hardware
will be doing the clipping for us.

v2: Rebase due to bigjoiner bitmask usage

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>