drm/i915: Split map_aux_ch() into per-platform arrays
The big switch+if statement mess in map_aux_ch() is
illegible. Split up into cleaner per-platform arrays
like we already have for the gmbus pins.

To save space in the arrays we shift away the lower nibble
of the VBT AUX CH byte since it's always zero anyway.

Extra care must be taken to not leave any holes in the
arrays because zero initialization would cause them to
be interpreted as AUX_CH_A. While technically we could
leave AUX_CH_NONE entries from the end of the array
(as the whole array would then be smaller) I decided to
fully pad out the arrays to set a good example.

TODO: Didn't bother with the platform variants beyond the
      ones that really need remapping, which means if the
      VBT is bogus we end up with a nonexistent aux ch.
      Might be nice to check this a bit better.
      Yet another bitmask in device info?

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>