drm/i915: Bump "early pixel count scaling" on ADL DP outputs
The spec is 110% obfuscated here. For some reason we should set
this to x4 on adl, but leave it to the default x2 on all other
platforms? HSD is useless.

Windows actually sets this to x4 when DSC is used, and leaves
it to x2 otherwise. And it does that on all platforms.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>