drm/i915/psr: Rewrite PSR2 selective fetch area calculations
The current PSR2 selective fetch area calculation code is
extremely confused about coordinate spaces. Rewrite it to
actually do the right thing.

The high level flow is as follows:
for each plane in the commit:
	1. collect the damage (if any) from the uapi state to sel_fetch.src
	2. translate the damage to clipped plane dst coordinaes in sel_fetch.dst
for each crtc in the commit:
	1. calculate the total damaged area on the crtc
	2. add all active planes to the state so we can calculate the
	   final selective fetch coordinates for each one
for each crtc in the commit:
	1. calculate the final selective fetch destination coordinates
	   for each plane
	2. calculate the final selective fetch source coordinates for
	   each plane

In the latter stages we may need to loop around a few times to
make sure we've considered issues such as plane scaling, cursor
planes, and plane source coordinate alignment. The final selective
fetch area will be expanded until all the requirements are met.

TODO: actually test it, and figure out where to put the code

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>