drm/i915: Update pipes in reverse order for bigjoiner
With bigjoiner the master crtc is the one that will send out the uapi
event/etc. We want that to happen after all the slaves are done. With
sync updates this generally doesn't matter as all the pipes will
complete via the imultanous blank interrupts, but with async flips
each pipe will take turns to do the async update, so doing the master
last makes it more likely that the slaves have already completed their
updates. Though it's still not 100% guaranteed as the latency for the
async flip can be different for each pipe (depends on FIFO fill levels
etc.). But this is certainly more likely to avoid userspace getting an
-EBUSY for the next async flip.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>