drm/atomi-helper: Allow the driver to customize drm_atomic_helper_setup_commit() behaviour
i915 will sometimes need to gang up multiple crtcs in order to
drive high resolution/dotclock modes. To facilitate that it
has a different idea of whether the crtc/plane is active than
what the drm core uapi state indicates. So we need a custom
mechanism to detemine that information for drm_atomic_helper_setup().
Add a set of function pointers that the driver can provide to
do that job.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>