drm/xe/guc: define PVC FW
Add PVC GuC FW definition. This is patch is for CI-only and it is meant
for the topic/xe-for-CI branch.

The new repo flow has the CI firmwares located under the intel-ci/
folder, so we need to update the path in the driver. This messes a bit
with the formatting, but since this is a CI-only patch it doesn't really
matter.

Link: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/1155
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Cc: John Harrison <John.C.Harrison@Intel.com>
Cc: Lucas De Marchi <lucas.demarchi@intel.com>
Acked-by: Gustavo Sousa <gustavo.sousa@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240308134202.901394-1-daniele.ceraolospurio@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>