drm/i915/dsb: Compute DEwake scanline more accurately
The current code calculating the DEwake scanline doesn't take
into account the fact that VRR might be getting disabled/enabled
during the commit. Take that into account to get a 100% correct
answer.

Technically this shouldn't matter as the final scanline
should be somewhere within vertical active (which won't change),
and so even the use of incorrect vtotal should cause an actual
change in the result. The only very minor difference might be
the extra scanline that the pre-ADL VRR hardware imposes to
the start of vblank. But one scanline here or there seems
rather inconsequential.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>