drm/i915/dmc: Reload PIPEDMC MMIO registers for pipe C/D on PTL+
On PTL+ the PIPEDMC apparently loses its MMIO state when PG2 goes down.
Reload the MMIO registers for the affected pipes when enabling the
PIPEDMC. So far I've not see this happen on PTL pipe A/B, nor on any
pipe on any other platform.

The DMC program RAM doesn't appear to need manual restoring, though
Windows appears to be doing exactly that on most platforms for some
of the pipes. None of this is really documented anywhere so it's all
just guesswork and empirical evidence it seems.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>